Sema: Split off Solution.h from ConstraintSystem.h
